split details into multiple files
backup/details has gotten to large to navigate.
This change splits up the file into multiple files
based on structure.
